Save scenarios for new projects
Previously, we would always use a "silent" save which would discard the
server's response in favor of the one available in the client. This was
useful since the client usually had newer values anyway, and autosaving
served the purpose of updating the server on changes from the client.

However, new scenarios are on exception where the server's response has
a newer value: the scenario's ID. In silent saves, we would discard the
ID along with everything else, and thus the client would treat scenario
as unsaved, causing new scenario generation on revisiting the project.

Thus, we switch to silently saving only pre-existing scenarios, and now
save new scenarios the regular way, ensuring the UI is updated when the
server's response is received.

Refs #849
Refs #739