Add basic routing
We add a new route, `/projects/`, which will take an authenticated user
to their project list. If the user is not logged in, we redirect them
to the home page.

In addition, we also redirect `/project/`, which is the project detail
route missing a project id, to the project list, instead of raising a
404.

We adjust tests to match, and add two new ones for the project listing.