Cache AoI and modification censuses.
* If a AoI census or modification censuses have already been generated
and are still valid, skip the geoprocessing tasks and run TR55 directly.
* If only the AoI census is available, generate the modification
censuses and provide the cached AoI census to TR-55.
* Modify the structure of the Scenario.census to contain the
modification hash from when the censuses were generated, as well
as both the AoI census and the modification censuses in the structure
expected in the geoprocessing tasks.
* Add new fields to Scenario to store the AoI census and modification
censuses. Scenario.census is no longer being used will be removed
in a separate migration.
* Update the client-side to handle the new Scenario fields.
* Update and add tests to test caching logic.

Refs #870