Add ProjectListingSerializer for My Projects
Previously, we would use the same ProjectSerializer for listing multiple
projects as we would for individual projects. This detailed result would
include complete polygons for Area of Interest, as well as all the data
for each scenario, resulting in a massive payload.

Since the My Projects view uses only a very small subset of information
provided, we create a new ProjectListingSerializer which quickly gives
only the required subset. We switch `GET /api/modeling/projects` to
respond with this new subset instead of the full listing, which is now
only employed for individual project calls.