Allow PATCH saving for Projects
Since the "My Projects" view only has a subset of a project's data, any
changes made from that page must augment, instead of replace, data on
the server. Thus, instead of using PUT, we must use PATCH.

We add a PATCH handler to the /api/modeling/projects/{id} endpoint, as
well as a `saveProjectListing` method for sending the PATCH request.