Allow sharing from My Projects view
The Share Modal prompts the user to make their project public prior to
sharing if it isn't already. Unfortunately, this prompt was very tied to
the old architecture: it expected an App.currentProject, and also would
saveProjectAndScenarios after changing privacy settings.

Since that doesn't really apply anymore, we switch the core modal to
instead look for a parameterized project, falling back in its absence to
App.currentProject, and saveProjectListing instead. This is sufficient
for both when saving from Modeling and My Projects views, because the
change applies only to the project and not any attached scenarios.

However, since now we are saving from two states: one where Project.user
is just an ID field (My Projects) and one where it is an entire object
(Modeling), but the PATCH expects it to be only an ID field, we ensure
that we send only an ID and not an entire object for PATCH requests.