Worker Affinity (2/2)
The parts of the the job chains which must remain on the same worker,
namely the top- and bottom-halves of the SJS submission/polling
functionality, are forced to remain on the same worker.  This is done so
that the polling task is not scheduled on another worker which does not
have access to the spark-jobserver on which the geoprocessing query has
been submitted.

Connects #886