Fill empty space when selecting Draw tools
Previously, if there was a location bar left over from the Analyze view,
and the user reset the AoI or selected a new draw tool, the location bar
would disappear but leave an empty white space.

2e57f36 tried to fix this but made the mistake of resetting the map only
after the user clicked "Reset", and not when the user simply selected a
different draw tool. By moving the call to `setDrawSize` within
`clearAoILayer`, we ensure that the map is always resized whenever the
AoI is cleared.

We switch from `setDrawSize(true)` to `setDrawSize(false)` because there
is no AoI to fit to, thus we don't attempt fitting, making the resize
transition slightly smoother.

Refs #893