Fix effective area calc for mods outside AoI.
* Set Modification.effectiveShape to an empty geometry when
the modification is outside the AoI. Previously, it was set
to undefined (turf.Intersect's return value when there is
no intersection), which resulted in effectiveArea not being
set initially because we abort calculating an area if there
is no shape. It appeared to be set correctly because when
a null value gets passed to Math.round, it returns 0.

Refs #846