Ensure sliders are visible in IE
`display: inline-block` and `width: auto` were incompatible together in
IE, so we give the precipitation slider an explicit width in the Model
view. In the Compare view, the container was not wide enough, so we make
it fill up all the available space to ensure that the slider is visible.

Vertical sliders cannot have any padding in IE.