Update UI after change for IE
IE doesn't trigger `input` event when the slider is being dragged, but
instead triggers `change` continuously. This is different from Firefox,
Chrome and Safari, which trigger `input` when dragging the slider, and
`change` when the user stops dragging.

Previously we would only update the UI on the `input` event to preview
the value selected by the user. Unfortunately, since IE never fires it,
the UI would not update even when the user stopped dragging. This edit
ensures that the UI is updated with the final value.

The actual fetching of the results is debounced, so even though we are
firing a lot more `change` events than were originally expected, there
are no more network requests and performance is quite consistent.