Enable screenshot in all cases
Previously we would only enable the screenshot library if we knew that
we were inside an embedded session. However, in certain edge cases with
login issues that failed to set the correct flags, this embed detection
would not work and screenshots would either show the ITSI login screen
or the MMW login screen. This would happen because the parent container
would ask for the screenshot but the app would not respond, since the
screenshot library hadn't been initialized. This would cause the server
to request the page itself, creating a new anonymous session and seeing
the login screens.

Now that the screenshot library is enabled in all cases, whenever there
is a request for a screenshot we send the full contents of the page and
get a proper response, even in the case of login failures and guest
sessions.