Always redirect after ITSI modal
Previously, we would only redirect to the "next" query param (or to home
if it was not specified) after the user continued registration. In cases
when registration was suspended or canceled, the application state would
be inconsistent and stuck.

By moving the redirection code to an overridden `onModalHidden` hook, we
ensure that the app always redirects regardless of the manner in which
it was closed.