Avoid username collisions with ITSI
Previously we would pre-populate the ITSI sign up modal with a username
taken from ITSI itself. Unfortunately, if the username already existed
in MMW, the user would get a "username already exists" error. While the
user was allowed to change the username and still be connected to ITSI
under the hood, this wasn't very apparent to many users, who would drop
out of the registration by clicking "Continue as Guest".

Now we add a `.itsi` suffix to their ITSI username, which significantly
lowers the chances of collision, while maintaining user friendliness.