Better handling of 404 projects
Previously, we would run into errors when cleaning up after a badly
loaded project view. Now, if a project fails to load, we clear the
global App.currentProject, as well as log a message to the console.
Additionally, while cleaning up, we check if App.currentProject is
defined, and only perform cleaning operations with it if it is.