Remove Chart Display Delay
When the user clicked the little "+" button to create a new scenario,
there was a slight delay between the time the spinners stop spinning and
the chart was displayed.

That was caused by the fact that the `reset` method on the
`ResultCollection` was firing a `reset` event which caused the spinner
to stop spinning.  Now, that `reset` call is made with the `silent`
option set to true so that the event is not fired, then it is later
manually fired after the model results are obtained.

Connects #940