Remove animation before taking screenshot
All four effect divs had fade animations on them. Unfortunately, when
shutterbug would send the HTML to the server, it would render and then
immediately take a screenshot before the effect divs could render.

By removing the animation classes before a screenshot is taken, we allow
the server to see the effect divs immediately, and have them show up in
the final image.

It's alright to not add these classes back since they are only used for
animating in. The animation afterwards when values change is controlled
by CSS transitions.