Shrink to fit in embed window
We reduce the minimum width to be 936px, which is the width of the embed
window in ITSI, instead of 1024px which is the width of landscape iPad.

We also switch from `col-md` to `col-xs` for the precipitation slider to
ensure that the three elements: min image, slider, and max image, are
always on the same line.

Refs #789