Always save scenarios silently, but update ID
This is a more correct implementation of #877. Previously, we
update the entire object with server values if it was new. This would
cause a number of local changes, such as `results.polling` which
controlled the spinners, to be overwritten by server defaults. Now we
only update the local object with fields that were not present locally
and are supplied by the server. Currently, these fields are `id`,
`created_at`, and `modified_at`, but the expression which simply updates
all fields not present locally will scale to include fields added in the
future.

Refs #940
Refs #876