Fix Modeling view screenshots
Because of the 5000px width set on .scenario-tabs-wrapper, and the way
PhantomJS sees things, some elements which are positioned absolutely and
have percent widths or `right: 0` would render way off screen.

We use the `shutterbug-saycheese` and `shutterbug-asyouwere` events
fired by shutterbug immediately before and after taking a screenshot to
add a fixed width, equivalent to the viewport's width, to
#model-output-wrapper, ensuring that all its contents are displayed
within visible bounds.