Disable 3D transforms in embed mode to fix screenshots
PhantomJS used for taking screenshots doesn't support CSS 3D transforms,
causing the map and drawn shapes to appear askew and displaced. The flag
`L_DISABLE_3D` instructs Leaflet not to use 3D transforms, but we cannot
abandon them entirely because of performance improvements they offer due
to hardware acceleration.

By disabling 3D transforms only for cases when we know we are running in
an embedded environment, we achieve a solution that helps in most cases.