Add popup to observation markers
Observation markers have a popup that is initially bound to the
measurement values that are included with the original metadata request.
When a popup is opened, that asset's recent values are fetched and the
UI is updated with those new values.  This appears to happen quickly,
but this separation allows the popup not to get bogged down by the
follow up call.