Upgrade SJS to 0.6.1; account for job status API change
Make use of the Spark Job Server (SJS) 0.6.1 container image, and
account for behavior changes on successful job competition. Before, when
a job completed successfully, we removed it from the list of jobs as a
maintenance task. Now, it appears that SJS takes care of this by itself.
In addition, it changed the status field in the job status API call from
"OK" to "FINISHED".

Attempts to resolve #1079.