Use backported kernel to avoid recent AUFS bugs
Bugs in a recent version of AUFS (included in the kernel) can produce
zombie processes that consume 100% CPU utilization. Using a backported
kernel allows things to continue to function as they did prior to the
bug. We'll remove this workaround once the fix is included in another
kernel update.

See also:

- https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1533043
- https://bugzilla.kernel.org/show_bug.cgi?id=109971