Integrate Sidebar.
Moves the analyze results into a tab on the modeling page. Note that
there are display and design issues with the chart and graphs that
will be taken care of in a future commit.

* Modify analyze views, templates, and styling so that they can
be used within a tab and more closely match design mock-ups.
* Shift modeling results from the footer to the sidebar.
* After drawing an area, navigate to the modeling page and take
care of analyze prep in prep for modeling.

Refs #1093