Load vizer layers asyncronously
Vizer Observation layers were requested async, but the layer control was
not rendered until their successful return.  This caused the control to
disappear when the Vizer API was down.  This commit updates the layer
control with vizer layers as they are returned, without waiting on them
to render.  If there are failures in loading the UI will report a
message on the observation tab.