Fix destroying of compare views.
* Because the CompareWindow view overwrote the destroy method
with a method that didn't destroy the view, when you navigated
back to the modeling page, the compare view remained open. This
was not exposed before because the modeling page reused the
same region as the compare view, so it overwrote the contents
of the compare view. To fix this, we rename CompareWindow.destroy
to CompareWindow.onDestroy, which will be called when the view
is destroyed.
* Make the footer region a regular region instead of a transition
region. No views that use the region are currently using a transition.