Fix units in analyze table header.
* In 2b39fe8, the AoI name and area were added to the analyze results.
Because this change set the shape attribute of the GeoModel for this
view, the model initialization logic overwrote the units
attribute. We manually calculated this value to match the units of
the analyze results.  When the table header was drawn, the units value
matched that of the AoI and not the result units. Here we send
along the manually calculated units as a separate variable to the
analyze table view. We also remove the line to manually set the units
of the GeoModel instance because the initialization logic will take
care of this for us.

Refs #1125