No longer POST JAR to SJS; ensure Upstart script respawns
In order to speed up the SJS initialization process, the job JAR no
longer needs to be POSTed to the SJS service. Instead, we take advantage
of the job-jar-paths configuration setting (added by @lossyrob).

In addition, the post-stop stanza in the Upstart job was updated to
ensure that a failure in the post-stop script does not prevent the
service from respawning on failure.