Add support for Ansible 2.0
Mostly removing a handful of deprecation warnings and upgrading to
versions of Ansible roles that work for Ansible 2.0. The role
specification itself has changed from a comma-delimited file to YAML.

An unexpected consequence of the upgrade revealed an issue Packer has
with uploading directories that contain symlinks. Previously, roles
pulled down by `ansible-galaxy` omitted the `examples` directory that
exists within all of our company-wide roles. Now, that directory is
included and Packer doesn't like it. Some code was introduced to the
Packer driver to remove the directories before Packer runs.