Reduce nesting, rotate parent
When we rotated only the slider (to be vertical instead of horizontal),
the parent div still extended horizontally and would capture clicks. By
reducing the amount of nesting in the markup, and rotating the parent
div instead, we ensure that the slider click handlers are confined to
its visible boundaries.