Preserve itsi_embed for new users
Previously, we would capture the itsi_embed flag from the initial
request via a query string, and store it in a session variable. This
works fine when dealing with returning users that are logging back in,
but fails for new users, since their session_id changes once they are
logged in as the new user. By extracting that value from the previous
session (if it existed) and populating the new session with it, we
ensure that the value is preserved.

This preserved value is used, among other things, for disabling 3D
transforms on the client side and allowing screenshots to work
correctly.