Preserve itsi_embed from login modal
Previously if a user were to log out while in an itsi_embed session, and
then use the login modal to login again, the new session would not have
had the itsi_embed flag. To preserve this information, we now use a
different login URL when logging in from an itsi_embed session, which
mimics the initial redirection when the app is loaded with the flag in
the URL set. This ensures that the new session also has itsi_embed set
correctly.

This ensures that, among other things, 3D transforms are disabled within
the itsi_embed environment, necessary for PhantomJS to take screenshots.