Display alert when RWD does not yield a valid watershed
There are certain conditions where the RWD job completes successfully,
but does not yield a valid watershed. Without this alert, there is no
indication that the job completed or not.

The reason this is a plain alert box instead of something nicer, is
because we intend to address the cause behind the null watersheds soon.

Ref: https://github.com/WikiWatershed/rapid-watershed-delineation/issues/27

Fixes #1166