Allow display of registration content
The previous map container was hidden during the transition to the
sidebar layout (c392ed6c75), but the registration content was not
updated to target the new container.  This restores the visibility of
the previous container, which is now only used for the account pages and
is empty in the main map mode.