Activate & deactivate layers on start without using location
This changes the `zoomToggle` method to update the active and inactive layers on the first (and only the first) detected `touchstart` and `mouseover` events. This ensures the active & inactive layers will get set even the map doesn't initially zoom in on a user's location.

Connects #1189
(configured Git user name & email, again!))