Add S3 backed persistent tile cache
In order to reduce the percentage of requests that cause tiles to be
rendered on the fly, introduce a lazily built persistent tile cache in
Amazon S3.

Tile requests used to be served in this fashion:

- Request to CloudFront is served by CloudFront edge cache
- Request to CloudFront passes through to origin (tiler ELB) and
response comes back through CloudFront

New tile request behavior:

- Request to CloudFront is...
  - Served by CloudFront edge cache (best case scenario)
  - Served by `tile-cache` S3 website, which is the CloudFront origin
  (second best case scenario)
  - Redirected to `https://color-tiles...` (tiler ELB) and client
  processes redirect

In the last scenario, 404s against the `tile-cache` S3 website cause the
redirects to `https://color-tiles...`. That endpoint is the tile ELB,
which is responsible for rendering a tile in the response and to S3
(using the AWS SDK for Node.js).