Add model package dropdown in Analyze view
There is now a dropdown to select the model package in the Analyze view. This is just a placeholder for a nicer dropdown that will be added after the prototype is done. The views shown for GWLF-E are just the TR-55 views for now, since it's not clear how they will be different. The backend does not return any results, so the views are blank.