Fix select model package in ITSI mode
This changes the behavior of selecting a model package when the itsi_embed flag is true, so that the user will be prompted if they select a different model package after returning from the model view, and will simply return to the same model if selecting the same model package. All other behavior should be the same as before.