Add MapShed SJS tasks
Given an operation name and input data, mapshed_start method will
look up geoprocessing service settings for that operation, hydrate
the template with the input data, and send that request. Then,
mapshed_finish method will look up the results, and will transform
the output of the service into a format that can easily be used.