Get streams from database
In order to overlay streams with a raster dataset such as NLCD, we
need to send the vector streams to the geoprocessing service. This
method finds all the streams contained within the polygon.

The result can contain LineStrings and MultiLineStrings. The query
also returns some empty GeometryCollections, which we exclude with
`NOT ST_IsEmpty`.