Add geop tasks to chain, derive AgLength
We add geop_tasks that given a geometry returns a group of Celery
chains, each corresponding to one combination of rasters and
vectors. Each chain is a sequence of mapshed_start, mapshed_finish
and some post-processing of the results.

The first task we add is nlcd_streams, which combines NLCD raster
with Streams vector data, and calculates the length of streams in
agricultural NLCD codes (81 and 82).

We also add a dummy identity task which alleviates an issue with
Celery which causes problems when a chord is the penultimate task
in a chain.