Celery Errors: More descriptive geop errors
If the geoprocessing service returns an error, we now include that
error message in the exception as well. This makes debugging them
considerably easier, since one no longer need to inspect Spark
JobServer with a particular Job ID to get the details. That can
still be done to see a full stack trace.