Celery Errors: Limited Chord Unlocks
In cases where one has a group of chains as a chord header, and
a task in the middle of a chain raises an Exception, the final
task in the chain would not execute and thus not be marked as
"ready", and thus the group would not be marked as "ready". Thus,
the chord would attempt to retry indefinitely.

Unfortunately, this indefinite retrial is hard-coded into Celery.
Thus, we shim the chord_unlock function with a variant that takes
the configurable CELERY_CHORD_UNLOCK_MAX_RETRIES variable. This
will ensure that chord unlocking would not proceed infinitely when
a chain in a group in the chord header fails.

This overridden method must be defined before a Celery instance is
created, thus the placing of the method at the beginning of the
file.