Celery Errors: Promptly mark jobs as failed
While the previous commits take care of failings of Celery, they
do not trigger the save_job_error handler. Thus, the job is not
marked as failed and the front-end keeps polling until it reaches
its own timeout.

Apparently, specifying a link_error to the final chain is not good
enough to catch errors in nested chains / chords / groups. Thus,
we edit our chain definition as follows:

 * start_gwlfe_job has been replaced by its contents, so there is
   one less level if indirection.
 * The error handler is also passed to geop_tasks, which in turn
   applies it to the last task in each chain. This ensures that
   as soon as one of the chains fail, the job is marked as such
   and the front-end can stop polling.
 * Earlier tasks in the geop chains handle exceptions, and forward
   them to the next task. It is the responsibility of the final
   task to inspect the result for the 'error' key, and raise an
   exception if found. By convention, the exception should be pre-
   fixed with the task's name for easy identification in the logs.