Celery Errors: Buffer handling chord errors
Attaching an error handler to a chord's body causes weird errors,
so we place a simple buffer between geop_tasks and collect_data,
making it not the chord body and thus handle errors correctly.

At first I used the simple identity function as the buffer, but
since we need to flatten the result of geop_tasks anyway, the
simple method was updated to do something useful instead.