Add GWLF-E BMP configuration
To dynamically generate the UI for each BMP, we need various pieces of information: which variables from the data model to show, which inputs to gather from the user, how to validate those inputs, and how to compute the output (the GWLF-E variables corresponding to BMPs). There is a lot of duplicated functionality across BMPs so I attempted to create lots of helper functions to generate these configuration objects. This architecture is more general than is needed in that it's compatible with multiple user input variables for a single BMP, which we don't have at the moment. I also added descriptions of some new BMPs to modificationConfig.json.