Generate BMP UIs
Using the configuration added in the last commit, generate the UI for each BMP. When clicking the Convert button, an object containing the output of the UI is printed to the console. The output consists of the BMP-related variables that will be used as input to GWLF-E when the model is run.