Force registration of MapShed tasks
Since MapShed tasks are defined in a submodule to a Django
app, and never explicitly specified in INSTALLED_APPS, the
autoloading functionality did not work reliably in all
environments. Particularly, it would work in local vagrant
setups, but not on the staging AWS server.

By including it in CELERY_IMPORTS we ensure that the tasks
are always loaded.