refactor: Reduce GWLF-E runs by hashing input, mods
For TR-55, we use the inputmod_hash field to judge whether the
results should be fetched again or not. If they don't match, then
we surmise that the results are stale and must be run again.

Previously, we didn't have an inputmod_hash field in the GWLF-E
endpoint. Thus, when the front-end would compare its value to the
locally computed one, they would always mismatch and the results
would always be reran. Now that we do incorporate the field, the
results should only rerun when the inputs / modifications are
actually different.