refactor: Extract Task View Model and Message View
This model was devised to display the loading and error values
in the Analyze phase. Since we need to do the same thing for when
gathering MapShed data, we extract this functionality into core
models and views so it can be reused. We also add the ability to
add a custom message to the loading stage.