Don't wait for Results to Analyze
Previously, we would wait for the data gathering and result
calculating steps to finish before showing the Analyze tab. This
was an extraneous wait, since the analysis is already available.

Now, we show the analysis results either immediately (when coming
from the Analyze tab) or in parallel (when reloading a scenario)
to the results.