Remember Analyze results between views
Now we will remember Analyze results not only when going from
Analyze to Model, but also from Model to Analyze. We clear any
saved results when leaving Draw, so that if users go all the
ways back to Draw and choose a different Area of Interest, that
actually triggers a legitimate re-analysis.