Export GMS from Scenario instead of Project
Since each scenario can contribute different values to the GMS
file given its modifications, we change to exporting from the
scenario instead of the project.

Furthermore, we now allow Current Conditions to have a dropdown
menu, since we need it to allow GMS Export. While this does allow
duplication of the "Share" functionality, since sharing the
Current Conditions scenario is the same as sharing the project,
consistency in menus is more important, and this duplication is
preferable to Current Conditions not having a Share option at all
when all other scenarios have it.