Restore sidebar collapse functionality
The sidebar collapse functionality would previously be
enabled by the pre-handler of the /project route. Since we
switched to using the /project/new route to initialie each
project for its chosen model package, the original handler
was no longer executing, and the button wasn't enabled.

By including that instruction in the new handler, we ensure
that the button is made visible. In addition, we push down
the map controls a bit so they don't overlap the button.