Show error message when drawn AoI too big
Since the amount of area we can analyze is limited by our
computational capacity, we disallow areas of interest above a
certain size. Previously we would simply cancel the drawing of a
very large polygon with no user feedback. Now they user is given
an error message when they draw one that is too big.