Show explicit timeout message
Previously both error and timeout conditions had the same messages
for the user. This was confusing and unhelpful, especially as the
most common case for error was a very large area of interest. By
including this in our response to the user, we make the app easier
to deal with and understand.