Ensure consistent ordering of layers
Previously, overlay layers were added in the map in the order that they
were selected from the layer control.  The boundaries being near the top
resulted in them often showing below the overlays, which would be
selected second.  This commit gives a consisten zIndex to each type of
layer so that the boundaries are always drawn on top.