Restrict DEP layers to PA Viewport
DEP Municipalities and Urban Areas control in the Layer Selector should
not be enabled unless the viewport contains PA.  The bounding geography
for PA was a rough sketch of just 10 vertices that covers the state, to
keep the file small and the Poly in Poly test fast.  Zoom level restrictions
on rendering still apply.