Replace all spaces with underscores
For friendlier file handling, we replace spaces with underscores
in both the project and scenario name before downloading it. The
previous implementation only replaced the first space inside the
string, rather than all of them. This is an unfortunate although
known quirk of JavaScript.

This solution replaces all occurrence of all whitespace within a
project or scenario's name with an underscore. Sourced from this
post in StackOverflow http://stackoverflow.com/a/3795130/2053314