Enable zoom/attribute dependent stream rendering
For both NHD Plus and DRB streams, configure the cartographic styling
to limit streams of a high Stream Order (scale is 10-0, large to small)
to be visible at low zoom levels (high altitude).  As the zoom level
increases, more streams of lower stream order are shown and the stroke
width of the rendered line corresponds to the relative stream order.
The effect should be a gradual increase in detail of streams as the
user zooms in, and streams of a higher Stream Order should appear
larger than those of a smaller order.

The intent isn't to make the stream order attribute apparent from
viewing the stream, but to give a relative indication of size and limit
the number of streams shown when zoomed out.

Since the DRB and NHD streams vary considerably in density of streams
at every zoom level, some of the sizes were adjusted to accomodate a
reasonable appearance for both datasets while keeping the width of the
stream the same, per stream order, per dataset.