Rename stream layers to invalidate cache
A temporary measure to allow style changes for these stream layers to
avoid having to wait for CloudFront distribution TTL to expire and browser
cache issues.  Each iteration of style changes will add an
additional cache busting characters to the name, so that changes can be
reviewed immediately from the client and are guaranteed not to be cached.