Separate, refactor analyze endpoints
Since all analyze endpoints follow the same pattern, we make a
generic implementation of start_celery_job which, given a list
of celery tasks, adds saving and error handling, and returns a
job id which can be used for polling.

We then refactor various analyze views to use this new generic
implementation, by giving it a specific job chain.