Show point source analysis results on map
Hovering over the point source location id in the point source analysis
results table will add a marker at the location of the point source, pan
the map to that location, and open a popup with the information from
the table for that source.

To make this possible, the latitude and longitude of each location are
returned in the analysis results (the state was also added too).

Some modifications were made to the popup template to make the text
cleaner in cases where there is no data for some of the header
information.

Refs #1445