Increase SJS JVM stack size to avoid stack overflow
The serializers Kryo provides use the call stack when serializing nested
objects. This change increases the JVM stack size to 5MB so that Kryo
serialization can squeeze through.

The stack size increase here is for all threads, so it is possible that
this change has a negative impact in other areas if the number of
threads grows large.

See: https://github.com/EsotericSoftware/kryo#stack-size