Draw RWD watersheds before polygon validation
This fixes a bug where RWD shapes do not appear on the map, because the
FeatureCollection returned by RWD throws an exception during shape
validation.

Fix this by converting the watershed from a FeatureCollection to a
MultiPolygon before validation. Also decided to draw the watershed
before the validation step just in case it fails.

The rationale is that we should always err on the side of displaying the
RWD watershed, even if it fails validation. Although the shape is always
displayed now, we only navigate to the analyze view if validation
is successful.

Connects #1521