Retry on Requests ConnectionError against SJS
When the Requests HTTP client is used to interact with Spark JobServer,
ensure that the Celery job is retried when a ConnectionError exception
is raised.

The retry() calls make use of the exc keyword argument. This is used to
supply a custom exception when the max restart limit is exceeded.