Turn off time-based caching on production
The production nginx configuration template makes static assets have a cache
that expires way in the future due to the `expires max;` line. When static
assets are updated, and the website is loaded without a refresh, the client
never bothers sending a request to see if the etag is out of date, and doesn't
get the update.