Fix AOI state issues
The areaOfInterest and areaOfInterestName pieces of state need to be synced
together so that the correct values are rendered in the footer.

Otherwise, it's possible to end up in a state where you have delineated a
watershed, but the footer displays "1 Square Km" (for example).

Another change was to ensure that the AOI footer always exists. This
fixes a bug where the footer would only appear for invalid shapes if
you had previously generated a valid shape, and then navigated back to
draw mode.