Add support for creating AoI via uploaded data
Adds support for creating AoIs from uploaded shapefiles and geojson.

The browser DataTransfer and FileReader APIs are used to upload and read
in the file.

At the moment, only the first feature from the uploaded file is used.
This may change in the future.

The UI for this feature is temporary. It will be moved once #1743
is implemented

Refs #1739