Activate tooltips on click instead of hover
Tooltips are now activated on click instead of hover so that links in
the tooltip text are now clickable. Previously, the tooltip closed as the
pointer was moved from the tooltip icon to the tooltip text, preventing
the user from clicking on any links in the text.

To better unify the UI/UX of the app, the tooltips in the legend were
updated to behave in the same way. Additionally, the style of both
tooltips was made consistent.

Links were added to the RWD tools. Also, a general link to documentation
about the various draw tools was added to the sidebar instead of each
draw tooltip (as previously requested). This seemed like a more appropriate
solution now that there is a space for explanatory text near the tools.

Refs #1749, #1233