Keep map within post-render bounds on load
On load, the map would expand to the full page width, and transition
to a reduced width after the application finishes loading. This led
to a jarring transition, especially when the application loaded
slowly.

By adding the classes that would be applied post-load to the .map-
container before it loads, we ensure it stays within the bounds of
actual rendering.

Home and Project views set the map at different sizes, so we choose
the right preset before loading to minimize post-load transitions.

Refs #1773 :lipstick: