Account for fat borders on touch devices
Touch devices add a 2px gray border to map controls. Since we
use box-sizing: border-box for the entire app, these borders
eat in to the controls' width rather than expanding out, causing
layout and display issues.

By updating the controls to be 4px wider and taller on touch
devices, we ensure that the controls are rendered correctly.

Refs #1782 :bug: :lipstick: