Cancel previous search request before making a new one
Fixes a bug where the loading icon would not always spin while requests
were loading. The promise chain from the previous search request was
putting the UI in a confusing state, so now we ensure that only 1
promise chain is active at a time by "cancelling" the previous search
request.