Increase number of results limit on boundary layer search
Instead of displaying 5 total results for all boundary layers, we now
show up to 3 results for each boundary layer.

Also, priority was renamed to rank. Since we now show results for each
boundary layer regardless of priority/rank, the only thing this number
does is determine the sort order for results on the frontend.